+++ include/tmc/all_headers.hpp
@@ -15,9 +15,9 @@
 #include "tmc/ex_braid.hpp"     // IWYU pragma: export
 #include "tmc/ex_cpu.hpp"       // IWYU pragma: export
 #include "tmc/external.hpp"     // IWYU pragma: export
+#include "tmc/spawn.hpp"        // IWYU pragma: export
 #include "tmc/spawn_func.hpp"   // IWYU pragma: export
 #include "tmc/spawn_many.hpp"   // IWYU pragma: export
-#include "tmc/spawn_task.hpp"   // IWYU pragma: export
 #include "tmc/spawn_tuple.hpp"  // IWYU pragma: export
 #include "tmc/sync.hpp"         // IWYU pragma: export
 #include "tmc/task.hpp"         // IWYU pragma: export
